+++ OpenMarket/MockURLSessionTests/MockURLSessionTests.swift
@@ -13,11 +13,11 @@ class MockURLSessionTests: XCTestCase {
     override func setUpWithError() throws {
         try super.setUpWithError()
     }
-
+    
     override func tearDownWithError() throws {
         try super.tearDownWithError()
     }
-
+    
     func test_receivePage_서버요청이_성공한경우에_받아온Json데이터가_MockData와같은지() {
         let mockURLSession = MockURLSession(isSuccess: true)
         let sut = URLSessionProvider(session: mockURLSession)
@@ -47,6 +47,20 @@ class MockURLSessionTests: XCTestCase {
                 XCTFail("서버 요청이 실패하지 않은 오류")
             case .failure(let error):
                 XCTAssertEqual(error, DataTaskError.incorrectResponseError)
+            }
+        }
+    }
+    
+    func test_receivePage_실제로서버요청을했을때_2번페이지데이터를_받아올수있는지() {
+        let sut = URLSessionProvider(session: URLSession.shared)
+        
+        sut.receivePage(number: 2, countOfItems: 10) { result in
+            switch result {
+            case .success(let data):
+                let responsedData = decodeMarket(type: Page.self, data: data)
+                XCTAssertEqual(responsedData?.totalCount, 325)
+            case .failure(_):
+                XCTFail("서버 데이터 불일치 오류")
             }
         }
     }
